Move symbol extension functions into `SymbolExt` (#2378)
